add tests for utils
