+++ tests/test_utils.py
@@ -5,6 +5,8 @@ from pathlib import Path
 from unittest.mock import patch
 import numpy as np
 import pytest
+import xarray as xr
+from test_datasets import data_folder
 from zampy.datasets import utils
 from zampy.datasets.dataset_protocol import SpatialBounds
 from zampy.datasets.dataset_protocol import TimeBounds
@@ -138,3 +140,107 @@ def test_time_bounds_to_year_month():
     expected = [("2010", "1")]
     year_month_pairs = utils.time_bounds_to_year_month(times)
     assert expected == year_month_pairs
+
+
+@pytest.fixture(scope="function")
+def dummy_dir(tmp_path_factory):
+    """Create a dummpy directory for testing."""
+    return tmp_path_factory.mktemp("data")
+
+
+def test_convert_to_zampy(dummy_dir):
+    """Test function for converting file to zampy format."""
+    ingest_folder = Path(data_folder, "era5")
+    utils.convert_to_zampy(
+        ingest_folder=Path(dummy_dir),
+        file=Path(ingest_folder, "era5_10m_v_component_of_wind_1996-1.nc"),
+        overwrite=True,
+    )
+
+    ds = xr.load_dataset(Path(dummy_dir, "era5_10m_v_component_of_wind_1996-1.nc"))
+
+    assert list(ds.data_vars)[0] == "northward_component_of_wind"
+
+
+def test_parse_nc_file_10m_wind():
+    """Test parsing netcdf file function with 10 meter velocity u/v component."""
+    variables = {
+        "10m_v_component_of_wind": "northward_component_of_wind",
+        "10m_u_component_of_wind": "eastward_component_of_wind",
+    }
+    for variable in variables:
+        ds = utils.parse_nc_file(data_folder / "era5" / f"era5_{variable}_1996-1.nc")
+        expected_var_name = variables[variable]
+        assert list(ds.data_vars)[0] == expected_var_name
+        assert ds[expected_var_name].attrs["units"] == "meter_per_second"
+
+
+def test_parse_nc_file_radiation():
+    """Test parsing netcdf file function with surface radiation."""
+    variables = {
+        "surface_thermal_radiation_downwards": "strd",
+        "surface_solar_radiation_downwards": "ssrd",
+    }
+    for variable in variables:
+        ds_original = xr.load_dataset(
+            data_folder / "era5" / f"era5_{variable}_1996-1.nc"
+        )
+        ds = utils.parse_nc_file(data_folder / "era5" / f"era5_{variable}_1996-1.nc")
+
+        assert list(ds.data_vars)[0] == variable
+        assert ds[variable].attrs["units"] == "watt_per_square_meter"
+        assert np.allclose(
+            ds_original[variables[variable]].values,
+            ds[variable].values * 3600,
+            equal_nan=True,
+        )
+
+
+def test_parse_nc_file_precipitation():
+    """Test parsing netcdf file function with precipitation."""
+    ds_original = xr.load_dataset(
+        data_folder / "era5" / "era5_mean_total_precipitation_rate_1996-1.nc"
+    )
+    ds = utils.parse_nc_file(
+        data_folder / "era5" / "era5_mean_total_precipitation_rate_1996-1.nc"
+    )
+    expected_var_name = "total_precipitation"
+
+    assert list(ds.data_vars)[0] == expected_var_name
+    assert ds["total_precipitation"].attrs["units"] == "millimeter_per_second"
+    assert np.allclose(
+        ds_original["mtpr"].values,
+        ds["total_precipitation"].values * utils.WATER_DENSITY / 1000,
+        equal_nan=True,
+    )
+
+
+def test_parse_nc_file_pressure():
+    """Test parsing netcdf file function with surface pressure."""
+    ds = utils.parse_nc_file(data_folder / "era5" / "era5_surface_pressure_1996-1.nc")
+    expected_var_name = "surface_pressure"
+
+    assert list(ds.data_vars)[0] == expected_var_name
+    assert ds["surface_pressure"].attrs["units"] == "pascal"
+
+
+def test_parse_nc_file_air_temperature():
+    """Test parsing netcdf file function with 2 meter temperature."""
+    ds = utils.parse_nc_file(
+        data_folder / "era5-land" / "era5-land_2m_temperature_1996-1.nc"
+    )
+    expected_var_name = "air_temperature"
+
+    assert list(ds.data_vars)[0] == expected_var_name
+    assert ds["air_temperature"].attrs["units"] == "kelvin"
+
+
+def test_parse_nc_file_dew_temperature():
+    """Test parsing netcdf file function with 2 meter dewpoint temperature."""
+    ds = utils.parse_nc_file(
+        data_folder / "era5-land" / "era5-land_2m_dewpoint_temperature_1996-1.nc"
+    )
+    expected_var_name = "dewpoint_temperature"
+
+    assert list(ds.data_vars)[0] == expected_var_name
+    assert ds["dewpoint_temperature"].attrs["units"] == "kelvin"
